backend/requirements.txt
Django~=2.2
djangorestframework~=3.13.1